refactor: :recycle: update helpers folder name to data
