+++ src/data/projects.ts
@@ -1,3 +1,5 @@
+// projects api data
+
 export const projects = [
   {
     id: 1,
