+++ src/pages/bookmarks/index.tsx
@@ -3,7 +3,7 @@ import { SegmentedControl, Title } from '@mantine/core'
 import useSWR from 'swr'
 import useWidth from '../../hooks/useWidth'
 import { getBookmarks } from '../../lib/raindrop'
-import { collections } from '../../helpers/collections'
+import { collections } from '../../data/collections'
 import Layout from '../../components/Layout/Layout'
 import Bookmark from '../../components/Bookmark/Bookmark'
 import CustomLoader from '../../components/CustomLoader/CustomLoader'
